% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperCode.R
\name{clean_hourly_data}
\alias{clean_hourly_data}
\title{Remove data measurements that are outside of credible ranges}
\usage{
clean_hourly_data(df, fast = TRUE, temp.min = -50, temp.max = 130,
  wind_spd.max = 80, wind_gust.max = 120, pressure.min = 20,
  pressure.max = 50, wind_chill.min = -50, wind_chill.max = 70,
  heat_index.min = 50, heat_index.max = 150, precip.max = 2)
}
\arguments{
\item{df}{dataframe with the weather data.}

\item{fast}{use first of duplicate station-time observations rather than mean()}

\item{temp.min}{numeric temperatures below this will be set to NA.}

\item{temp.max}{numeric temperatures above this will be set to NA.}

\item{wind_spd.max}{numeric wind speeds above this will be set to NA.}

\item{wind_gust.max}{numeric wind speeds above this will be set to NA.}

\item{pressure.min}{numeric pressure below this will be set to NA.}

\item{pressure.max}{numeric pressure above this will be set to NA.}

\item{wind_chill.min}{numeric wind chill below this will be set to NA.}

\item{wind_chill.max}{numeric wind chill above this will be set to NA.}

\item{heat_index.min}{numeric heat index below this will be set to NA.}

\item{heat_index.max}{numeric heat index above this will be set to NA.}

\item{precip.max}{numeric precip above this will be set to NA.}
}
\value{
dataframe with extreme values set to NA
}
\description{
\code{clean_hourly_data} Filter out extreme and implausible values from retrieved
  weather data.  Most limits are user parametrizable, although precip and wind speeds
  are rejected if nonpositive, and visibility and humidity are restricted to be 
  between 0 and 100.
}
\seealso{
\code{\link{update_weatherdata}} \code{\link{rebuild_weather}}
}
