% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveCode.R
\name{merge_location_data}
\alias{merge_location_data}
\title{Merge two dataframes of location data, keeping most recently scraped data}
\usage{
merge_location_data(df1, df2, byvar = "localdate", rankvar = "fetchtime",
  sortvar = "date")
}
\arguments{
\item{df1}{data frame of archived wunderground data.}

\item{df2}{data frame of archived wunderground data.}

\item{byvar}{variable to group by in choosing data from df1 and df2.}

\item{rankvar}{variable to use in choosing data from df1 and df2.}

\item{sortvar}{variable to sort results by.}
}
\value{
a dataframe containing data from df1 and df2, where data is 
  included rom the dataframe which has the largest rankvar if the 
  same byvar group is present in both data frames, or if it is only in 
  one of df1 or df2.
}
\description{
\code{merge_location_data} Given two data frames, group observations, then 
  select which dataset to use based on rank.
}
\seealso{
\code{\link{update_weather_data}}    \code{\link{store_weather_data}}
}
