% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairCode.R
\name{report_weather}
\alias{report_weather}
\title{Summarize a weatherdata pair}
\usage{
report_weather(weatherpair)
}
\arguments{
\item{weatherpair}{list of periodic and daily weather data, as produced by 
\code{\link{rebuild_weather}}}
}
\description{
\code{report_weather} Summarize the number of dates in the weather data archive, 
     and share of dates covered for each station.
}
\seealso{
\code{\link{update_weather}} \code{\link{rebuild_weather}}
}
