% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperCode.R
\name{update_weatherdata}
\alias{update_weatherdata}
\title{Update weatherdata pair with new data from wunderground}
\usage{
update_weatherdata(weatherpair, weatherstation, begin.date = NA,
  end.date = NA, update.dir = "", ...)
}
\arguments{
\item{weatherpair}{a 2 item list containing the hourly and daily dataframes}

\item{weatherstation}{string containing the station ID (preface zip codes 
with ZIPxxxxx)}

\item{begin.date}{a string "YYYYMMDD" identifying the first date to fetch in
the interval.  If not specified, it chooses the last date specied for the
weatherstation in th weatherpair dataframes.  If there is no such date, the
begin.date is set to the same value as end.date.}

\item{end.date}{a string "YYYMMDD" identifying the last date to fetch in the 
interval.  If not specified, the current Sys.Date() is used.}

\item{update.dir}{string containing the locally-archived wunderground data 
updates directory, so that data fetched can be re-used.}

\item{...}{validity limit variables for data cleaning}
}
\value{
a list containing two dataframes.  One contains approximately hourly measures
  of temperature, precipitation, wind, etc, along with time and weatherstation id. 
  The other contains daily mins, max, means and totals for the things in the hourly 
  reports.  Mins and maxs are based on all daily observations plus interpolated 
  values for midnight (local time) at the beginning and end of the day.  Means are
  based on data plus the midnight interpolations, weighted by the elapsed time between
  them.  If the data is for a time and place where Daylight Savings Time is in effect, 
  there will be one 23 hour and one 25 hour day per year.  Precip totals are based on
  the average hourly rainfall rate and a 24 hour day.
}
\description{
\code{update_weatherdata} Fetch the necessary data from wunderground,
  archive it, process it and update the hourly detail and the daily 
  summary data frames.
}
\details{
The data archive and update files are stored in a dataframe as returned
from wunderground.  Timezones are set to local for the station, so when 
you combine these into a single frame, the TZ info will be wiped out, 
and the date/time variables will be affected.

Before combining these dataframes from different stations, we store the Timezone,
and the local Time and Date as character strings reflecting the time at the site, 
regardless of your tz settings.
}
\seealso{
\code{\link{rebuild_weather}}     \code{\link{clean_hourly_data}}
}
