% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairCode.R
\name{weather_interval}
\alias{weather_interval}
\title{Return a dataframe of weather variables characterizing an interval}
\usage{
weather_interval(weatherdata, weatherlocation, localdateBegin, localtimeBegin,
  durationHours, dataWindow = 120)
}
\arguments{
\item{weatherdata}{data frame containing weather data}

\item{weatherlocation}{char vector containing the station IDs (preface
zip codes with ZIPxxxxx)}

\item{localdateBegin}{char vector containing the local date of the beginning
of the interval "YYYYMMDD"}

\item{localtimeBegin}{char vector containing the local time of the beginning
of the interval "HHMMSS"}

\item{durationHours}{numeric length of the interval in hours}

\item{dataWindow}{number of minutes past ends of interval to look for data}
}
\value{
a dataframe containg the following variables:
}
\description{
\code{weather_interval} Return a dataframe of weather variables 
  characterizing an interval.
}
\details{
If the interval begins or end in the hour where daylight saving
causes the hour to repeat (fall back), it will be assumed that it 
begins or ends on the first pass.
}
\seealso{
\code{\link{update_weather}}     \code{\link{clean_hourly_data}}
}
