% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairCode.R
\name{weather_report}
\alias{weather_report}
\title{Summarize a weatherdata pair}
\usage{
weather_report(weatherpair)
}
\arguments{
\item{weatherpair}{list of periodic and daily weather data, as produced by 
\code{\link{rebuild_weather}}}
}
\description{
\code{weather_report} Summarize the number of dates in the weather data archive, 
     and share of dates covered for each station.
}
\seealso{
\code{\link{update_weatherdata}} \code{\link{rebuild_weather}}
}
