% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wunderWrapper.R
\name{wunderWrapper}
\alias{wunderWrapper}
\title{wunderWrapper: A package to generate historic data hourly requests to the package 
  rwunderground and save the results in .rda files.  
  
wunderWrapper offers a few functions to retrieve, save and format historical weather
  data from wunderground.com.  You will need a key from wunderground.com, which 
  has limits on request volume and rate.  See rwunderground package documentation 
  if you want to do more.}
\description{
wunderWrapper: A package to generate historic data hourly requests to the package 
  rwunderground and save the results in .rda files.  
  
wunderWrapper offers a few functions to retrieve, save and format historical weather
  data from wunderground.com.  You will need a key from wunderground.com, which 
  has limits on request volume and rate.  See rwunderground package documentation 
  if you want to do more.
}
\section{fetch data from wunderground.com}{

  \link{update_weatherdata}, \link{nocap_build_archive}
}

\section{rebuild hourly/daily dataframe pair from archive data}{

  \link{rebuild_weather}
}

