% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperCode.R
\name{rebuild_weather}
\alias{rebuild_weather}
\title{Pull together the weather archive files and stored updates}
\usage{
rebuild_weather(archive.dir, update.dir = "")
}
\arguments{
\item{archive.dir}{string containing the locally-archived wunderground data directory.}

\item{update.dir}{string containing the locally-archived wunderground data updates directory.}
}
\value{
A list containing two dataframes.  One contains approximately hourly measures
  of temperature, precipitation, wind, etc, along with time and weatherstation id. 
  Th other contains daily mins, max, and totals for the things in the hourly reports.
}
\description{
\code{rebuild_weather} Read in all of the archived and stored updates for  
 weather data.  It returns a list of 2 dataframes (hourly and daily).
}
\details{
The data archive and update files are stored in a dataframe as returned
from wunderground.  Timezones are set to local, so when you combine these
into a single frame, the TZ info will be wiped out, and the date/time
variables will be affected.

Before combining these dataframes from different stations, we store the Timezone,
and the local Time and Date as character strings reflecting the time at the site, regardless
your tz settings.
}
\seealso{
\code{\link{update_weatherdata}}
}
